Compiler warning...
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@2832 7a7537e8-13f0-0310-91df-b6672ffda945

+++ filter/rastertohp.c
@@ -1,5 +1,5 @@
 /*
- * "$Id: rastertohp.c,v 1.19 2002/03/01 19:53:36 mike Exp $"
+ * "$Id: rastertohp.c,v 1.20 2002/09/26 15:16:07 mike Exp $"
  *
  *   Hewlett-Packard Page Control Language filter for the Common UNIX
  *   Printing System (CUPS).
@@ -561,7 +561,7 @@ CompressData(unsigned char *line,	/* I - Data to compress */
   * Set the length of the data and write a raster plane...
   */
 
-  printf("\033*b%d%c", line_end - line_ptr, plane);
+  printf("\033*b%d%c", (int)(line_end - line_ptr), plane);
   fwrite(line_ptr, line_end - line_ptr, 1, stdout);
 }
 
@@ -804,5 +804,5 @@ main(int  argc,		/* I - Number of command-line arguments */
 
 
 /*
- * End of "$Id: rastertohp.c,v 1.19 2002/03/01 19:53:36 mike Exp $".
+ * End of "$Id: rastertohp.c,v 1.20 2002/09/26 15:16:07 mike Exp $".
  */
